training_tools/requirements.txt
torch
torchvision
transformers
pillow
tqdm
